requirements.txt
opencv-python
numpy
streamlit
pyttsx3
mediapipe
Pillow
deepface
torch
tf-keras
